requirements.txt
numpy
scipy
pandas
matplotlib
temporaldata
torch
lightning
hydra-core
einops
scikit-learn
torcheval

# [dev]
ipykernel
debugpy